pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["dados <felipe_ludson@hotmail.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
structlog = "^24.1.0"
pydantic = "^2.6.1"
requests = "^2.31.0"
httpx = "^0.26.0"
typer = {extras = ["all"], version = "^0.9.0"}
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
fastapi = "^0.115.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"
